fixed 69f95b9 from https://gitee.com/zhangxiao72/ace_ace_engine/pulls/1729 Ace CodeDex 看板告警清零

+++ frameworks/core/components_v2/list/render_list.cpp
@@ -1507,8 +1507,7 @@ void RenderList::CreateDragDropRecognizer()
                     Point point = info.GetGlobalPoint() - targetRenderlist->GetGlobalOffset();
                     auto newListItem = targetRenderlist->FindCurrentListItem(point);
                     if (static_cast<int32_t>(targetRenderlist->GetIndexByListItem(newListItem)) > -1) {
-                        renderList->insertItemIndex_ =
-                            static_cast<int32_t>(targetRenderlist->GetIndexByListItem(newListItem));
+                        renderList->insertItemIndex_ = targetRenderlist->GetIndexByListItem(newListItem);
                     }
                     if (targetRenderlist == renderList) {
                         (targetRenderlist->GetOnItemDragMove())(dragInfo,
